requirements.txt
streamlit>=1.29.0
pandas>=2.1.3
plotly>=5.18.0
python-dotenv>=1.0.0
openai>=1.3.0
streamlit-option-menu>=0.3.6
numpy>=1.24.0
matplotlib>=3.7.0
streamlit-option-menu

